Adding python file and requirements file to github
This is the python file with the code for my streamlit app that i wrote in VS code. I am uploading it here so that I can connect it to the streamlit cloud and have a record of it here. I am also uploading the requirements.txt file to ensure other users or my future self can run this code succesfully.